chore: move the undo button to where the other input actions are
